requirements.txt
async-generator==1.10
attrs==22.1.0
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
filelock==3.7.1
h11==0.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.7.0
itemloaders==1.0.4
jmespath==1.0.1
lxml==4.9.1
outcome==1.2.0
parsel==1.6.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
PyJWT==2.4.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
selenium==4.4.3
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tldextract==3.3.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
twilio==7.12.1
Twisted==22.4.0
typing_extensions==4.3.0
urllib3==1.26.11
w3lib==1.22.0
webdriver-manager==3.8.3
wsproto==1.1.0
zope.interface==5.4.0
